Add help metadata to task row components
- Added detailed help metadata to the points badge
- Added help metadata to the edit button
- Added help metadata to the task title area
- Added help metadata to the critical task indicator
- Added help metadata to the completion slider with usage hints
- Each component now provides helpful context when tapped in help mode
- Included comprehensive usage hints for all task row elements

+++ Points/TaskManager.swift
@@ -120,7 +120,7 @@ class TaskManager {
         task.points = points
         task.target = target
         task.completed = 0
-        task.max = max ?? Constants.Defaults.taskMax // Use provided max or default constant
+        task.max = max ?? Int16(Constants.Defaults.taskMax) // Use provided max or default constant
         task.position = Int16(fetchTasks().count)
         task.routine = routine
         task.optional = optional
